Implemented Qnet Slicing and found issue with net
The volume neural network's ouput ranges from less than 0 to greater than 1, this is an issue and integrating in it can produce negative totals. need to retrain network sooner than expected. in mean time, will add 1 to output and divide by 2, after integration (over -1 to 1) this is add 2 to the integral and divide by 2